\name{fraser}
\alias{fraser}
\title{
Fraser River at Hope, mean monthly flow (cms),  1912.3-1991.12
}
\description{
Mean monthly flow data.
}
\usage{data(fraser)}
\format{A ts object containing monthly data for 1912.3-1991.12}
\source{Environment Canada CDROM}
\keyword{datasets}
% Converted by Sd2Rd version 1.21.
