\name{ozone}
\alias{ozone}
\title{
Ozone concentration, downtown L.A., 1955.1-1972.12
}
\description{
Ozone concentration, downtown L.A., 1955.1-1972.12
in parts per hundred million
}
\usage{data(ozone)}
\format{A ts object containing monthly data for 1955.1-1972.12.}
\source{}
\keyword{datasets}
% Converted by Sd2Rd version 1.21.
